llm-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sentence-transformers==2.5.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
numpy==1.26.4
pydantic==2.6.3